update the interface return type
update the interface return type

+++ cgs-product/cgs-policy/src/main/java/com/coral/cgs/service/IUserService.java
@@ -7,5 +7,5 @@ import com.coral.cgs.model.User;
  */
 public interface IUserService {
 
-    int insertUser(User user);
+    Long insertUser(User user);
 }
